Add RequestScopeDeterminator for service init
Otherwise some event listeners will perform despite the fact that they
should NOT be run in either backend or frontend.

NOTE: when implementing services it is NOT enough to check the "current"
scope in initialization (as done in previous implementation) but it is
needed to check the scope of the "current" request at execution time.

The requests might differ as we may be running in a sub request which
changed scope.

Long term all of this should be refactored to proper populator and
builder chains (listeners/subscribers should become tagged services).